% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_easylinear.R
\name{all_easylinear}
\alias{all_easylinear}
\title{Easy Growth Rates Fit to data Frame}
\usage{
all_easylinear(df, criteria, time = "time", y = "value", h = 5,
  quota = 0.95)
}
\arguments{
\item{df}{data frame of observational data}

\item{criteria}{vector of criteria defining subsets in the data frame}

\item{time}{character vectors with name independent variable}

\item{y}{character vector with name of dependent variable}

\item{h}{with of the window (number of data)}

\item{quota}{part of window fits considered for the overall linear fit (relative to max. growth rate)}
}
\value{
list with parameters of all  fits
}
\description{
Determine maximum growth rates from log-linear part of the growth curve for
a series of experiments.
}
\examples{
library("growthrates")
L <- all_easylinear(bactgrowth,
                     criteria=c("strain", "conc", "replicate"),
                     time="time", y="value")
summary(L)
coef(L)
rsquared(L)

results <- results(L)

library(lattice)
xyplot(mu ~ conc|strain, data=results)
}
\references{
Hall, B. G., H. Acar and M. Barlow (2013). Growth Rates Made Easy.
  Mol. Biol. Evol. 31: 232-238 doi:10.1093/molbev/mst197
}
\seealso{
Other fitting.functions: \code{\link{all_growthmodels}};
  \code{\link{all_splines}}; \code{\link{fit_easylinear}};
  \code{\link{fit_growthmodel}}; \code{\link{fit_spline}}
}

