% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{as.Date.time_period}
\alias{as.Date.time_period}
\alias{as.POSIXct.time_period}
\title{Convert time period to dates}
\usage{
\method{as.Date}{time_period}(x, ...)

\method{as.POSIXct}{time_period}(x, ...)
}
\arguments{
\item{x}{a time_period}

\item{...}{not used}
}
\value{
a vector of dates representing the start of each of the input
\code{time_period} entries
}
\description{
Convert time period to dates
}
\section{Functions}{
\itemize{
\item \code{as.POSIXct(time_period)}: Convert to a vector of POSIXct

}}
