% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{as.timeseries}
\alias{as.timeseries}
\alias{is.timeseries}
\alias{is.multinomial_ts}
\title{A time-series is a regular sequence of times with other data}
\usage{
as.timeseries(df, ...)

is.timeseries(df)

is.multinomial_ts(df)
}
\arguments{
\item{df}{a dataframe with \code{time} or \code{date} and optionally \code{class} or class
defined by grouping}

\item{...}{
  Arguments passed on to \code{\link[=as.timed_df]{as.timed_df}}
  \describe{
    \item{\code{days_in_period}}{the number of days in one time unit. May be non
integer. If the time column in the dataframe is already a time_period then
this is optional, but if provided will recalibrate the time_period to a new
interval. Otherwise it is required}
    \item{\code{start_date}}{the zero day. If the time column in the dataframe is
already a time_period then this is optional, but if provided will
recalibrate the time_period to a new interval. If the dataframe has a
date column then this is optional and the zero day is assumed to be the
start of the timeseries.}
    \item{\code{class_col}}{the column name of a class column if present. Defaults to
"class" but will only be used if the column is present. The class column
may be a binomial or multinomial class marker.}
    \item{\code{grp_label}}{a glue specification to apply to the groups of the \code{df}
dataframe to create a unique label for the groups.}
  }}
}
\value{
a timeseries object with groupwise unique time and date columns
}
\description{
The time series is expected to have at least one of a \code{time} or \code{date}
columns and any other set of data columns. There can be multiple time-series
within this if the \code{class} column is present or the input is grouped. For
every \code{class} and \code{time}/\code{date} combination there can be only one entry.
Missing values are not filled in. If there is grouping (but no \code{class}
column) then it is temporarily stored and replaced with an integer \code{class}
column. This can be restored with \code{restore_grouping()} later
}
\section{Functions}{
\itemize{
\item \code{is.timeseries()}: Check is a timeseries

\item \code{is.multinomial_ts()}: Check is a mulinomial timeseries

}}
\examples{
library(tidyverse)
library(growthrates)

 tibble(time = 0:99, value=rnorm(100)) \%>\%
   as.timeseries(days_in_period=7,start_date="2020-01-01")

 # defining a multinomial with a class column
 tmp1 = tibble(
     class = c(rep(1,100),rep(2,100)),
     time = rep(0:99,2),
     value = rnorm(200)
  ) \%>\% as.timeseries(days_in_period=7,start_date="2020-01-01")

 restore_groups(tmp1)
 # still a timeseries:
 class(restore_groups(tmp1))


 # defining a multinomial with a grouing
 tmp = tibble(
     mygroup = c(rep("a",100),rep("b",100)),
     time = rep(0:99,2),
     value = rnorm(200)
   ) \%>\%
   group_by(mygroup) \%>\%
   as.timeseries(days_in_period=7,start_date="2020-01-01")

 restore_groups(tmp)
 # no longer a time series:
 class(restore_groups(tmp))
}
