% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{breaks_log1p}
\alias{breaks_log1p}
\title{A scales breaks generator for log1p scales}
\usage{
breaks_log1p(n = 5, base = 10)
}
\arguments{
\item{n}{the number of breaks}

\item{base}{the base for the breaks}
}
\value{
a function for ggplot scale breaks
}
\description{
A scales breaks generator for log1p scales
}
\examples{
library(tidyverse)
ggplot(diamonds, aes(x=price))+
  geom_density()+
  scale_x_continuous(trans="log1p", breaks=ggrrr::breaks_log1p)
}
