% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cost.R
\name{cost}
\alias{cost}
\title{Cost Function for Nonlinear Fits}
\usage{
cost(p, obs, FUN, ...)
}
\arguments{
\item{p}{initial parameter vector of the growth model}

\item{FUN}{function of growth model to be fitted}

\item{\dots}{additional parameters passed to the optimizer}
}
\value{
an object of class \code{modCost}, see \code{\link{modCost}} in
  package \pkg{FME}
}
\description{
Defines a cost function from the residual sum of squares between model
and observational data.
}
\keyword{internal}

