% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{covid_infectivity_profile}
\alias{covid_infectivity_profile}
\title{The covid_infectivity_profile dataframe structure specification}
\format{
A dataframe containing the following columns:
\itemize{
\item boot (anything) - a bootstrap identifier
\item time (positive_double) - the end of the time period (in days)
\item probability (proportion) - the probability of infection between previous time period until \code{time}
}

Must be grouped by: boot (exactly).

A default value is defined.
}
\description{
The covid_infectivity_profile dataframe structure specification
}
\concept{datasets}
\keyword{interfaces}
