% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{date_to_time}
\alias{date_to_time}
\title{Convert a set of dates to numeric timepoints}
\usage{
date_to_time(
  dates,
  unit = .day_interval(dates),
  start_date = getOption("day_zero", "2019-12-29")
)
}
\arguments{
\item{dates}{a vector of dates to convert}

\item{unit}{a specification of the unit of the resulting time series.
Will be determined from periodicity of dates if not specified. If another
\code{time_period} is given as the unit then the}

\item{start_date}{the origin of the conversion. Defaults to the beginning of the COVID pandemic}
}
\value{
a vector of class \code{time_period}
}
\description{
Using a start_date and a unit specification
}
\examples{
times = date_to_time(as.Date("2019-12-29")+0:100, "1 week")
dates = time_to_date(times)
}
\concept{time_period}
