% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubling-time.R
\name{doubling_time}
\alias{doubling_time}
\title{Doubling time from growth rate}
\usage{
doubling_time(x, ...)
}
\arguments{
\item{x}{a dataframe calculated from either proportion or incidence growth
rate calculations:

e.g. A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

No mandatory groupings.

No default value.

OR

A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
\item relative.growth.fit (double) - an estimate of the relative growth rate
\item relative.growth.se.fit (double) - the standard error the relative growth rate
\item relative.growth.0.025 (double) - lower confidence limit of the relative growth rate
\item relative.growth.0.5 (double) - median estimate of the relative growth rate
\item relative.growth.0.975 (double) - upper confidence limit of the relative growth rate
}

No mandatory groupings.

No default value.}

\item{...}{not used}
}
\value{
the same dataframe with additional columns for doubling time
or relative doubling time plus confidence intervals.
}
\description{
The unit of doubling times is always days.
}
\examples{
growthrates::england_covid \%>\%
  growthrates::poisson_locfit_model(window=21) \%>\%
  growthrates::doubling_time() \%>\%
  dplyr::glimpse()


}
