% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_consensus_growth_rate}
\alias{england_consensus_growth_rate}
\title{The SPI-M-O England consensus growth rate}
\format{
A dataframe containing the following columns:
\itemize{
\item date (date) - the date of the estimate
\item low (numeric) - the lower published estimate of the growth rate
\item high (numeric) - the higher published estimate of the growth rate
}

No mandatory groupings.

No default value.

111 rows and 3 columns
}
\usage{
data(england_consensus_growth_rate)
}
\description{
SPI-M-O used a range of different statistical and mechanistic models to
produce estimates of the growth rate of the epidemic from various data
sources (including with an early version of \code{growthrates}).
}
\concept{datasets}
\keyword{datasets}
