% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_consensus_rt}
\alias{england_consensus_rt}
\title{The england_consensus_rt dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item date (as.Date) - the date column
\item low (numeric) - the low column
\item high (numeric) - the high column
}

No mandatory groupings.

No default value.

113 rows and 3 columns
}
\usage{
data(england_consensus_rt)
}
\description{
The england_consensus_rt dataset
}
\keyword{datasets}
