% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_covid}
\alias{england_covid}
\title{Daily COVID-19 case counts by age group in England}
\format{
A dataframe containing the following columns:
\itemize{
\item date (as.Date) - the date column
\item class (enum(\verb{00_04},\verb{05_09},\verb{10_14},\verb{15_19},\verb{20_24},\verb{25_29},\verb{30_34},\verb{35_39},\verb{40_44},\verb{45_49},\verb{50_54},\verb{55_59},\verb{60_64},\verb{65_69},\verb{70_74},\verb{75_79},\verb{80_84},\verb{85_89},\verb{90+})) - the class column
\item count (numeric) - the count column
\item denom (numeric) - the denom column
\item time (as.time_period) - the time column
}

Must be grouped by: class (and other groupings allowed).

No default value.

26790 rows and 5 columns
}
\usage{
data(england_covid)
}
\description{
A dataset of the daily count of covid cases by age group in England
downloaded from the UKHSA coronavirus API, and formatted for
use in \code{growthrates}. A denominator is calculated which is the overall
positive count for all age groups. This data set can be used to calculate
group-wise incidence and absolute growth rates and group wise proportions and
relative growth rates.
}
\keyword{datasets}
