% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_covid_proportion}
\alias{england_covid_proportion}
\title{England COVID by age group for ascertainment}
\format{
A dataframe containing the following columns:
\itemize{
\item class (character) - the age group
\item date (date) - the start date of a week
\item count (numeric) - the count of COVID positives
\item denom (numeric) - the number of COVID tests performed
\item population (numeric) - the size of the population at this age group
\item time (as.time_period) - the time column (weekly)
}

Must be grouped by: class (and other groupings allowed).

No default value.

1050 rows and 6 columns
}
\usage{
data(england_covid_proportion)
}
\description{
An age group stratified dataset from
}
\details{
\itemize{
\item the coronavirus.gov.uk site for
positive cases aggregated to 10 year age groups and by weekly time.
\item NHS test
and trace date which reported regional by age group testing effort aggregated
to country level.
\item ONS 2021 census population aggregated to 10 year age
groups.
}
}
\keyword{datasets}
