% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_demographics}
\alias{england_demographics}
\title{England demographics}
\format{
A dataframe containing the following columns:
\itemize{
\item class (enum(\verb{00_04},\verb{05_09},\verb{10_14},\verb{15_19},\verb{20_24},\verb{25_29},\verb{30_34},\verb{35_39},\verb{40_44},\verb{45_49},\verb{50_54},\verb{55_59},\verb{60_64},\verb{65_69},\verb{70_74},\verb{75_79},\verb{80_84},\verb{85_89},\verb{90+})) - the class column
\item population (numeric) - the population column
\item baseline_proportion (numeric) - the baseline_proportion column
}

Must be grouped by: class (and other groupings allowed).

No default value.

19 rows and 3 columns
}
\source{
https://www.ons.gov.uk/file?uri=/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationandhouseholdestimatesenglandandwalescensus2021/census2021/census2021firstresultsenglandwales1.xlsx
}
\usage{
data(england_demographics)
}
\description{
Population counts by 5 year age group for England only from the 2021 census.
}
\keyword{datasets}
