% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_events}
\alias{england_events}
\title{The england_events dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item label (character) - the label column
\item start (date) - the start column
\item end (date) - the end column
}

No mandatory groupings.

No default value.

13 rows and 3 columns
}
\usage{
data(england_events)
}
\description{
The england_events dataset
}
\keyword{datasets}
