% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_nhs_app}
\alias{england_nhs_app}
\title{NHS COVID-19 app data}
\format{
A dataframe containing the following columns:
\itemize{
\item date (date) - the start date of the week
\item alerts (integer) - the count of self-isolation alerts
\item visits (integer) - the number of venue check-ins representing visits to
social venues.
\item time (as.time_period) - the time column
}

No mandatory groupings.

No default value.

137 rows and 4 columns
}
\usage{
data(england_nhs_app)
}
\description{
check-in (social activity) and alerts (self isolation instruction) data from
the NHS COVID-19 app, aggregated to country level on a week by week basis.
}
\keyword{datasets}
