% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_variants}
\alias{england_variants}
\title{The england_variants dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item date (date) - the date column
\item time (as.time_period) - the time column
\item class (enum(\code{Other},\code{Alpha (B.1.1.7)},\code{Delta (B.1.617.2)},\code{Delta (AY.4)},\code{Omicron (Other)},\code{Omicron (BA.2)},\code{Omicron (BA.4)},\code{Omicron (BA.5)},\code{XBB (Other)},\code{Kraken (XBB.1.5)},\code{Arcturus (XBB.1.16)},\code{Eris (EG.5.1)})) - the class column
\item who_class (enum(\code{Other},\code{Alpha},\code{Delta},\code{Omicron},\code{Kraken},\code{Arcturus},\code{Eris})) - the who_class column
\item count (numeric) - the count column
\item denom (numeric) - the denom column
}

Must be grouped by: class (and other groupings allowed).

No default value.

479 rows and 6 columns
}
\usage{
data(england_variants)
}
\description{
The england_variants dataset
}
\keyword{datasets}
