% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite-estimators.R
\name{estimate_binomial_proportion}
\alias{estimate_binomial_proportion}
\title{A binomial proportion estimate and associated exponential growth rate}
\usage{
estimate_binomial_proportion(
  timeseries,
  ...,
  window = 14,
  deg = 2,
  quick = FALSE
)
}
\arguments{
\item{timeseries}{a timeseries, which must have a \code{time} and one combination of
\code{proportion} OR \code{count} and \code{total} OR \code{class} and \code{count}. Optionally it may
have a \code{group} column which is treated as independent timeseries,}

\item{...}{not used. Allows use in a group modify}

\item{window}{the number fo days data to include, Should be upped for weekly
time series.}

\item{deg}{the polynomial degree (higher = more wiggly)}

\item{quick}{should rely solely on count data wherever possible.}
}
\value{
a timeseries including \code{proportion} columns, and \code{relative.growth}
columns. The growth is relative to whatever end up in the \code{total} column.
This might be the total number of tests (which can be complex to interpret
in the situation where testing is changing), or might be the total number
of all variants of the class (e.g. when comparing genomic variants).
}
\description{
takes a list of times and counts based on and fits a
quasi-binomial model fitted with a logit link function to proportion data
using local regression. Behaviour depends on inputs a bit.
}
\details{
This expects timeseries to contain one combination of:
\itemize{
\item time and proportion columns (+/- class), and in this case fits
independent models to any classes and with proportion as the response and
time as the predictor. This format assumes you have pre-calculated proportions
for each day, but does not take into account sample size for individual days
when calculating confidence intervals.
\item time and count and total columns - Total for example might be all tests
conducted for example. In this case if \code{quick=TRUE} it calculates
the proportion and proceeds as above. If \code{quick=FALSE} it expands the
time-series of counts to a line list of individuals with either 1 for
positive or 0 for negative outcome (count positives, total-count negatives).
The model is then fitted to individuals with a outcome predicted by time. This
is only worth doing if numbers are small. If class is present, then each
class will be treated as an independent time series.
\item time and class and count, without total columns. In this case the class
columns are treated as a complete set of observations, and the algorithm
calculates a total based on the sum of the class counts. This results in a one
versus others comparison, which is only strictly valid is class is binomial
but will work for multinomial. Once the total is defined the algorithm
proceeds as above depending on the \code{quick} parameter.
}
}
\examples{
if (FALSE) {
  utils::vignette("estimators-example", package="growthrates")
}
}
