% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite-estimators.R
\name{estimate_multinomial_proportion}
\alias{estimate_multinomial_proportion}
\title{Estimate the probabilities of a multinomial class based on count data}
\usage{
estimate_multinomial_proportion(timeseries, deg = 2, ...)
}
\arguments{
\item{timeseries}{a timeseries of \code{class} and \code{count} data where class has
2 or more levels. The classwise counts are used to assess a time varying
probability using spline terms.}

\item{deg}{the degree of the polynomial spline}

\item{...}{
  Arguments passed on to \code{\link[splines:bs]{splines::bs}}
  \describe{
    \item{\code{x}}{the predictor variable.  Missing values are allowed.}
    \item{\code{df}}{degrees of freedom; one can specify \code{df} rather than
    \code{knots}; \code{bs()} then chooses \code{df-degree} (minus one
    if there is an intercept) knots at suitable quantiles of \code{x}
    (which will ignore missing values).  The default, \code{NULL},
    takes the number of inner knots as \code{length(knots)}.  If that is
    zero as per default, that corresponds to \code{df = degree - intercept}.}
    \item{\code{knots}}{the \emph{internal} breakpoints that define the
    spline.  The default is \code{NULL}, which results in a basis for
    ordinary polynomial regression.  Typical values are the mean or
    median for one knot, quantiles for more knots.  See also
    \code{Boundary.knots}.}
    \item{\code{degree}}{degree of the piecewise polynomial---default is \code{3} for
    cubic splines.}
    \item{\code{intercept}}{if \code{TRUE}, an intercept is included in the
    basis; default is \code{FALSE}.}
    \item{\code{Boundary.knots}}{boundary points at which to anchor the B-spline
    basis (default the range of the non-\code{\link{NA}} data).  If both
    \code{knots} and \code{Boundary.knots} are supplied, the basis
    parameters do not depend on \code{x}.  Data can extend beyond
    \code{Boundary.knots}.}
  }}
}
\value{
a timeseries of \code{class} and \code{proportion} columns where the proportion
is an estimate of the multinomial class probability at any given time.
}
\description{
takes a list of times, classes and counts, e.g. a COGUK variant like data set
with time, (multinomial) class (e.g. variant) and count being the count in
that time period. Fits a quadratic B-spline on time to the proportion of the
data using \code{nnet::multinom}.
}
\examples{
if (FALSE) {
  utils::vignette("variant-proportions", package="growthrates")
}
}
