% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite-estimators.R
\name{estimate_poisson_rate}
\alias{estimate_poisson_rate}
\title{Estimate an incidence rate from count data using a poisson regression}
\usage{
estimate_poisson_rate(
  timeseries,
  window = 14,
  deg = 2,
  daily_incidence = FALSE
)
}
\arguments{
\item{timeseries}{a timeseries including a \code{count} column. Any \code{group} or \code{class}
columsn are used for grouping and treated as independent time series}

\item{window}{the width of the window of data to choose in days. If your data
is weekly you probably will want to up this from it default value of 14.
Larger numbers give fewer wiggles.}

\item{deg}{the polynomial degree to fit (higher numbers give more wiggles)}

\item{daily_incidence}{Do you want to return estimates of the incidence as a daily
rate regardless of the periodicity of the input, or stick with the time unit of
the input? regardless of this setting results will be returned as a daily
frequency (but might be in cases per week if the input data was weekly).}
}
\value{
a new timeseries with time, date, and various incidence and growth
rate metrics
}
\description{
takes a list of times and counts based on a quasi-poisson model fitted to
count data using local regression. expects timed_df to contain time and count
columns only.
}
\examples{
if (FALSE) {
  utils::vignette("estimators-example", package="growthrates")
}
}
