% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extcoef_logistic.R
\name{extcoef_logistic}
\alias{extcoef_logistic}
\title{Extended Results of a Logistic Growth Curve}
\usage{
extcoef_logistic(object, quantile = 0.95, time = NULL, ...)
}
\arguments{
\item{object}{model object fited by fit_growthmodel}

\item{quantile}{value relative to $K$ fr the quantile method}

\item{time}{2-valued vector of the search interval for the independent
variable (time).
Set this manually if saturation is not reached within the
observation time that is taken from the data.}

\item{...}{reserved for future extensions}
}
\value{
vector that contains the fitted parameters and some
  derived characteristics (extended parameters) of the logistic
  function.
}
\description{
Estimate model-specific derived parameters of the logistic growth
model
}
\details{
This function returns the estimated parameters of a logistic growth model
 (y0, mumax, K) and a series of estimates for the time of approximate saturation.
 The estimates are defined as follows:
 \itemize{
   \item turnpoint: time of turnpoint (50\% saturation)
   \item sat_deriv2: time of the minimum of the 2nd derivative
   \item sat_mumax: time of the intercept between the tangent at mumax and the capacity limit K
   \item sat_quantile: time when a quantile of K (default 0.95) is reached
 }
 The code and naming of the parameters is preliminary and experimental, and
   may change in future versions.
}
\examples{


}
