% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{fdmy}
\alias{fdmy}
\title{Format date as dmy}
\usage{
fdmy(date)
}
\arguments{
\item{date}{a date to convert}
}
\value{
the formatted date
}
\description{
Format date as dmy
}
\examples{
fdmy(Sys.Date())
}
