% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_easylinear.R
\name{fit_easylinear}
\alias{fit_easylinear}
\title{Fit Exponential Growth Model with a Heuristic Linear Method}
\usage{
fit_easylinear(x, y, h = 5, quota = 0.95)
}
\arguments{
\item{x}{vector of independend variable (e.g. time)}

\item{y}{vector of dependend variable (concentration of organisms)}

\item{h}{with of the window (number of data)}

\item{quota}{part of window fits considered for the overall linear fit 
(relative to max. growth rate)}
}
\value{
list with parameters of the fit
}
\description{
Determine maximum growth rates from log-linear part of the growth curve using
a heuristic similar to the ``growth rates made easy''-method of Hall et al. (2013).
}
\examples{
data(bactgrowth)

splitted.data <- multisplit(bactgrowth, c("strain", "conc", "replicate"))
dat <- splitted.data[[1]]

plot(value ~ time, data=dat)
fit <- fit_easylinear(dat$time, dat$value)

plot(fit)
plot(fit, log="y")
plot(fit, which="diagnostics")

fitx <- fit_easylinear(dat$time, dat$value, h=8, quota=0.95)

plot(fit, log="y")
lines(fitx, pch="+", col="blue")

plot(fit)
lines(fitx, pch="+", col="blue")


}
\references{
Hall, B. G., H. Acar and M. Barlow (2013). Growth Rates Made Easy. 
  Mol. Biol. Evol. 31: 232-238 doi:10.1093/molbev/mst197
}
\seealso{
Other fitting functions: \code{\link{all_easylinear}},
  \code{\link{all_growthmodels}},
  \code{\link{all_splines}}, \code{\link{fit_growthmodel}},
  \code{\link{fit_spline}}
}

