% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{full_seq.numeric}
\alias{full_seq.numeric}
\title{Create the full sequence of values in a vector}
\usage{
\method{full_seq}{numeric}(x, period = 1, tol = 1e-06, ...)
}
\arguments{
\item{x}{a numeric or date vector}

\item{period}{Gap between each observation. The existing data will be checked to ensure that it is actually of this periodicity.}

\item{tol}{Numerical tolerance for checking periodicity.}

\item{...}{for subtype methods}
}
\value{
a vector of the same type as the input
}
\description{
This is useful if you want to fill in missing values that should have been observed but weren't. For example, full_seq(c(1, 2, 4, 6), 1) will return 1:6.
}
\examples{
full_seq(c(1, 2, 4, 5, 10), 1)
}
