% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_growthmodel-class.R
\name{function_growthmodel-class}
\alias{function_growthmodel-class}
\title{Union Class of Growth Model or Function}
\description{
This class defines formal requirements for package and user-provided functions
  that can be used as models describing time-dependent growth of organisms.
}

