% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{germany_covid}
\alias{germany_covid}
\title{Weekly COVID-19 case counts by age group in Germany}
\format{
A dataframe containing the following columns:
\itemize{
\item class (enum(\verb{0–14},\verb{15–19},\verb{20–24},\verb{25–29},\verb{30–39},\verb{40–49},\verb{50–59},\verb{60–69},\verb{70–79},\verb{80+},\code{Unknown}, .ordered=TRUE)) - the class column
\item date (as.Date) - the date column
\item count (integer) - the count column
\item time (as.time_period) - the time column
\item denom (integer) - the denom column
}

Must be grouped by: class (and other groupings allowed).

No default value.

2070 rows and 6 columns
}
\usage{
data(germany_covid)
}
\description{
A dataset of the weekly count of covid cases by age group in Germany
downloaded from the Robert Koch Institute Survstat service, and formatted for
use in growth rates. A denominator is calculated which is the overall
positive count for all age groups. This data set can be used to calculate
group-wise incidence and absolute growth rates and group wise proportions and
relative growth rates.
}
\keyword{datasets}
