% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{germany_demographics}
\alias{germany_demographics}
\title{Germany demographics}
\format{
A dataframe containing the following columns:
\itemize{
\item class (enum(\verb{0–14},\verb{15–19},\verb{20–24},\verb{25–29},\verb{30–39},\verb{40–49},\verb{50–59},\verb{60–69},\verb{70–79},\verb{80+}, .ordered=TRUE)) - the class column
\item population (integer) - the population column
}

Must be grouped by: class (and other groupings allowed).

No default value.

10 rows and 2 columns
}
\usage{
data(germany_demographics)
}
\description{
Derived from the Robert Koch Survstat service by comparing counts and
incidence rates.
}
\keyword{datasets}
