% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow_baranyi.R
\docType{data}
\name{grow_baranyi}
\alias{grow_baranyi}
\title{The Baranyi and Roberts Growth Model}
\format{\preformatted{function (time, parms)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 43 17 52 1 17 1 43 52
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000000b47b900> 
 - attr(*, "pnames")= chr [1:4] "y0" "mu" "K" "h0"
 - attr(*, "class")= chr "growthmodel"
}}
\usage{
grow_baranyi
}
\arguments{
\item{time}{vector of time steps (independend variable)}

\item{parms}{named parameter vector of the logistic growth model with:
\itemize{
  \item \code{y0} initial value of abundance
  \item \code{mu} maximum growth rate (1/time)
  \item \code{K} carrying capacity (max. abundance)
  \item \code{h0} parameter specifying the initial physiological state of
    organisms (e.g. cells) and in consequence the lag phase
    (h0 = max growth rate * lag phase).
}}
}
\value{
vector of dependend variable (\code{y})
}
\description{
The growth model of Baranyi and Roberts (1995) written as analytical solution
of the system of differential equations.
}
\examples{
time <- seq(0, 30, length=200)
y    <- grow_baranyi(time, c(y0=0.01, mu=.5, K=0.1, h0=5))[,"y"]
plot(time, y, type="l")
plot(time, y, type="l", log="y")
}
\references{
Baranyi, J. and Roberts, T. A. (1994).
A dynamic approach to predicting bacterial growth in food.
International Journal of Food Microbiology, 23, 277-294.

Baranyi, J. and Roberts, T.A. (1995). Mathematics of predictive microbiology.
International Journal of Food Microbiology, 26, 199-218.
}
\seealso{
Other growth models: \code{\link{grow_exponential}};
  \code{\link{grow_genlogistic}},
  \code{\link{grow_genlogistic.R}},
  \code{\link{ode_genlogistic}};
  \code{\link{grow_gompertz}}; \code{\link{grow_huang}};
  \code{\link{grow_logistic}}; \code{\link{grow_richards}};
  \code{\link{grow_twostep}}, \code{\link{grow_twostep.R}},
  \code{\link{ode_twostep}}
}
\keyword{datasets}

