% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow_gompertz.R
\name{grow_gompertz}
\alias{grow_expgompertz}
\alias{grow_gompertz}
\title{Growth Models According to Gompertz}
\usage{
grow_gompertz(time, parms)

grow_expgompertz(time, parms)
}
\arguments{
\item{time}{vector of time steps (independend variable)}

\item{parms}{named parameter vector of the logistic growth model with:
\itemize{
  \item \code{mu} intrinsic growth rate (1/time)
  \item \code{K} carrying capacity (max. total concentration of cells)
  \item \code{lambda} parameter specifying the lag-phase
  \item \code{alpha} scaling parameter
  \item \code{tshift} shifting parameter

}}
}
\value{
vector of dependend variable (\code{y})
}
\description{
Gompertz growth model written as analytical solution of the differential
  equation system and an expanded version that is more flexible.
}
\examples{
time <- seq(0, 30, length=200)
y    <- grow_gompertz(time, c(mu=1, K=10, lambda=5))[,"y"]
plot(time, y, type="l", ylim=c(0, 20))

y    <- grow_expgompertz(time, c(mu=2, K=10, lambda=2, alpha=.1, tshift=30))[,"y"]
lines(time, y, col="blue")
}
\seealso{
Other growth models: \code{\link{grow_exponential}};
  \code{\link{grow_logistic}}; \code{\link{grow_richards}};
  \code{\link{grow_twostep}}, \code{\link{ode_twostep}}
}

