% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow_huang.R
\name{grow_huang}
\alias{grow_huang}
\title{Growth Model According to Huang}
\usage{
grow_huang(time, parms)
}
\arguments{
\item{time}{vector of time steps (independend variable)}

\item{parms}{named parameter vector of the logistic growth model with:
\itemize{
  \item \code{y0} initial value of population measure
  \item \code{mu} maximum growth rate (1/time)
  \item \code{K} carrying capacity (max. total concentration of cells)
  \item \code{alpha} shape parameter determining the curvature
  \item \code{lambda} parameter determining the lag time

}}
}
\value{
vector of dependend variable (\code{y})
}
\description{
Huangs growth model ......
}
\details{
The naming of parameters used here follows the convention of Tsoularis 2001
but uses \code{mu} for growtrate and \code{y} for abundance to make them
more compatible with other growth functions.
}
\examples{
time <- seq(0, 30, length=200)
y    <- grow_huang(time, c(y0=0.01, mu=.1, K=0.1, alpha=1.5, lambda=3))[,"y"]
plot(time, y, type="l")
plot(time, y, type="l", log="y")
}
\references{
Huang, L. (2011) A new mechanistic growth model for simultaneous
determination of lag phase duration and exponential growth rate and a new
Belehdradek-type model for evaluating the effect of temperature on growth rate.
Food Microbiology , 2011, 28, 770 - 776
}
\seealso{
Other growth models: \code{\link{grow_baranyi}};
  \code{\link{grow_exponential}};
  \code{\link{grow_genlogistic}},
  \code{\link{grow_genlogistic.R}},
  \code{\link{ode_genlogistic}};
  \code{\link{grow_gompertz}}; \code{\link{grow_logistic}};
  \code{\link{grow_richards}}; \code{\link{grow_twostep}},
  \code{\link{grow_twostep.R}}, \code{\link{ode_twostep}}
}

