% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow_logistic.R
\name{grow_logistic}
\alias{grow_logistic}
\title{Logistic Growth Model}
\usage{
grow_logistic(time, parms)
}
\arguments{
\item{time}{vector of time steps (independend variable)}

\item{parms}{named parameter vector of the logistic growth model with:
\itemize{
  \item \code{y0} initial value of population measure
  \item \code{mu} intrinsic growth rate (1/time)
  \item \code{K} carrying capacity (max. total concentration of cells)
}}
}
\value{
vector of dependend variable (\code{y})
}
\description{
Classical logistic growth model written as analytical solution of the differential equation.
}
\examples{
time <- seq(0, 30, length=200)
y    <- grow_logistic(time, c(y0=1, mu=0.5, K=10))[,"y"]
plot(time, y, type="l")
}
\seealso{
Other growth models: \code{\link{grow_baranyi}};
  \code{\link{grow_exponential}};
  \code{\link{grow_genlogistic}},
  \code{\link{grow_genlogistic.R}},
  \code{\link{ode_genlogistic}};
  \code{\link{grow_gompertz}}; \code{\link{grow_huang}};
  \code{\link{grow_richards}}; \code{\link{grow_twostep}},
  \code{\link{grow_twostep.R}}, \code{\link{ode_twostep}}
}

