% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow_richards.R
\name{grow_richards}
\alias{grow_richards}
\title{Growth Model According to Richards}
\usage{
grow_richards(time, parms)
}
\arguments{
\item{time}{vector of time steps (independend variable)}

\item{parms}{named parameter vector of the logistic growth model with:
\itemize{
  \item \code{mu} growth rate (..... different interpretation cfrom exp growth .....)
  \item \code{K} carrying capacity (max. total concentration of cells)
  \item \code{lambda} parameter specifying the lag-phase
  \item \code{shape} shape parameter determining the curvature

}}
}
\value{
vector of dependend variable (\code{y})
}
\description{
Richards growth model written as analytical solution of the differential equation.
}
\examples{
time <- seq(0, 30, length=200)
y    <- grow_richards(time, c(mu=1, K=10, lambda=5, shape=2))[,"y"]
plot(time, y, type="l")
y    <- grow_richards(time, c(mu=1, K=10, lambda=2, shape=100))[,"y"]
lines(time, y, col="red")
y    <- grow_richards(time, c(mu=1, K=10, lambda=0, shape=.2))[,"y"]
lines(time, y, col="blue")
}
\seealso{
Other growth models: \code{\link{grow_expgompertz}},
  \code{\link{grow_gompertz}};
  \code{\link{grow_exponential}};
  \code{\link{grow_logistic}}; \code{\link{grow_twostep}},
  \code{\link{ode_twostep}}
}

