% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow_genlogistic.R, R/grow_twostep.R
\name{ode_genlogistic}
\alias{grow_genlogistic}
\alias{grow_twostep}
\alias{ode_genlogistic}
\alias{ode_twostep}
\title{Generalized Logistic Growth Model}
\usage{
ode_genlogistic(time, y, parms, ...)

grow_genlogistic(time, parms, ...)

ode_twostep(time, y, parms, ...)

grow_twostep(time, parms, ...)
}
\arguments{
\item{time}{vector of simulation time steps}

\item{y}{named vector with initial value of the system (e.g. cell concentration)}

\item{parms}{parameters of the growth model
\itemize{
   \item \code{mu} intrinsic growth rate (1/time)
   \item \code{K} carrying capacity (max. total concentration of cells)
   \item \code{alpha, beta, gamma} parameters determining the shape of growth.
     Setting all values to one returns the ordinary logistic function.
}}

\item{\dots}{placeholder for additional parameters (for user-extended versions of this function)}

\item{time}{vector of simulation time steps}

\item{y}{named vector with initial values of the system (yi, ya: concentration of inactive resp. active cells)}

\item{parms}{parameters of the growth model
\itemize{
   \item \code{kw} activation (``wakeup'') constant (1/time)
   \item \code{mu}  intrinsic growth rate (1/time)
   \item \code{K}  carrying capacity (max. total concentration of cells)
}}

\item{\dots}{placeholder for additional parameters (for user-extended versions of this function)}
}
\value{
matrix of class deSolve containing the simulation outputs

\itemize{
\item \code{time} time of the simulation
\item \code{y} total cell concentration
\item \code{log_y} natural log of total cell concentration
}

matrix of class deSolve containing the simulation outputs

\itemize{
\item \code{time} time of the simulation
\item \code{yi} concentration of inactive cells
\item \code{ya} concentration of active cells
\item \code{y} total cell concentration
\item \code{log_y} natural log of total cell concentration
}
}
\description{
Generalized logistic growth model solved as differential equation.

System of two differential equations describing bacterial growth as two-step process of activation (or adaptation) and growth.
}
\details{
The generalized logistic according to Tsoularis (2001) is an extremely flexible
  model that covers exponential and logistic growth, Richards, Gompertz, von
  Bertalanffy, and some more as special cases.

  The differential equation is solved numerically, where function
  \code{ode_genlogistic} is the differential equation, while
  \code{grow_genlogistic} runs a numerical simulation over time.

Function \code{ode_twostep} is the system of differential equations, whereas
  \code{grow_twostep} runs a numerical simulation over time.
}
\examples{
time <- seq(0, 30, length=200)
parms <- c(mu=0.5, K=10, alpha=1, beta=1, gamma=1)
y0    <-  c(y=.1)
out   <- ode(y0, time, ode_genlogistic, parms)
plot(out)

out2 <- ode(y0, time, ode_genlogistic, parms = c(mu=0.2, K=10, alpha=2, beta=1, gamma=1))
out3 <- ode(y0, time, ode_genlogistic, parms = c(mu=0.2, K=10, alpha=1, beta=2, gamma=1))
out4 <- ode(y0, time, ode_genlogistic, parms = c(mu=0.2, K=10, alpha=1, beta=1, gamma=2))
out5 <- ode(y0, time, ode_genlogistic, parms = c(mu=0.2, K=10, alpha=.5, beta=1, gamma=1))
out6 <- ode(y0, time, ode_genlogistic, parms = c(mu=0.2, K=10, alpha=1, beta=.5, gamma=1))
out7 <- ode(y0, time, ode_genlogistic, parms = c(mu=0.3, K=10, alpha=1, beta=1, gamma=.5))
plot(out, out2, out3, out4, out5, out6, out7)

plot(ode(y0, time, ode_genlogistic, parms = c(mu=1, K=10, alpha=2, beta=.8, gamma=5)))
time <- seq(0, 30, length=200)
parms <- c(kw = 0.1,	mu=0.2, K=0.1)
y0    <-  c(yi=0.01, ya=0.0)
out   <- ode(y0, time, ode_twostep, parms)

plot(out)
}
\references{
Tsoularis, A. (2001) Analysis of Logistic Growth Models.
Res. Lett. Inf. Math. Sci, (2001) 2, 23-46.
}
\seealso{
Other growth.models: \code{\link{grow_exponential}};
  \code{\link{grow_gompertz}}; \code{\link{grow_logistic}};
  \code{\link{grow_richards}}

Other growth.models: \code{\link{grow_exponential}};
  \code{\link{grow_gompertz}}; \code{\link{grow_logistic}};
  \code{\link{grow_richards}}
}

