% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_growthmodel-class.R
\name{growthmodel-class}
\alias{growthmodel-class}
\title{Class of Growth Model Functions}
\description{
This class is used for the parametric \code{grow_\dots} functions of the
package and can also be used for user-defined functions to describe
time-dependent growth of organisms.
}
\seealso{
the constructor function \code{\link{growthmodel}} how to create
  instances of this class.
}

