% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{is.Date}
\alias{is.Date}
\title{Check whether vector is a date}
\usage{
is.Date(x)
}
\arguments{
\item{x}{a vector to check}
}
\value{
TRUE if dates, FALSE otherwise
}
\description{
Check whether vector is a date
}
\examples{
is.Date(Sys.Date())
}
\concept{time_period}
