% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{labels.time_period}
\alias{labels.time_period}
\title{Label a time period}
\usage{
\method{labels}{time_period}(
  object,
  ...,
  dfmt = "\%d/\%b",
  ifmt = "{start} — {end}",
  na.value = "Unknown"
)
}
\arguments{
\item{object}{a set of decimal times as a time_period}

\item{...}{not used}

\item{dfmt}{a \code{strptime} format specification for the format of the date}

\item{ifmt}{a glue spec referring to \code{start} and \code{end} of the period as a
formatted date}

\item{na.value}{a label for \code{NA} times}
}
\value{
a set of character labels for the time
}
\description{
Create a set of labels for a time period based on the start and duration of
the period. The format is configurable using the start and end dates and the
\code{dfmt} and \code{ifmt} parameters, however if the time period has names then these
are used in preference.
}
\examples{
eg = as.time_period(Sys.Date()+0:10*7, anchor="start")
labels(eg)
labels(eg, ifmt="{start}", dfmt="\%d/\%b/\%y")
labels(eg, ifmt="until {end}", dfmt="\%d \%b \%Y")

# labels retained in constructor:
eg2 = Sys.Date()+0:10*7
names(eg2) = paste0("week ",0:10)
labels(eg2)
labels(as.time_period(eg2, anchor="start"))
}
\concept{time_period}
