% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{logit_trans}
\alias{logit_trans}
\title{logit scale}
\usage{
logit_trans()
}
\value{
A scales object
}
\description{
it perform logit scaling with right axis formatting. To not be used directly but with ggplot (e.g. ggplot2::scale_y_continuous(trans = "logit") )
}
\examples{

library(ggplot2)
library(tibble)

tibble::tibble(pvalue = c(0.001, 0.05, 0.1), fold_change = 1:3) \%>\%
 ggplot2::ggplot(aes(fold_change , pvalue)) +
 ggplot2::geom_point() +
 ggplot2::scale_y_continuous(trans = "logit")

}
