% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{max_date}
\alias{max_date}
\title{The maximum of a set of dates}
\usage{
max_date(x, ...)
}
\arguments{
\item{x}{a vector of dates}

\item{...}{ignored}
}
\value{
a date. `0001-01-01`` if there is no well defined minimum.
}
\description{
\code{max.Date} returns an integer and \code{-Inf} for a set of \code{NA} dates. This
is usually inconvenient.
}
\examples{
max_date(NA)
}
