% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{min_date}
\alias{min_date}
\title{The minimum of a set of dates}
\usage{
min_date(x, ...)
}
\arguments{
\item{x}{a vector of dates}

\item{...}{ignored}
}
\value{
a date. \code{9999-12-31} if there is no well defined minimum.
}
\description{
\code{min.Date} returns an integer and \code{Inf} for a set of \code{NA} dates. This
is usually inconvenient.
}
\examples{
min_date(NA)
}
