% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_set_generics.R, R/multisplit.R
\docType{methods}
\name{multisplit}
\alias{multisplit}
\alias{multisplit,data.frame,character-method}
\alias{multisplit,data.frame,formula-method}
\alias{parse_formula}
\title{Split Data Frame According to Multiple groups}
\usage{
multisplit(data, grouping, ...)

parse_formula(grouping)

\S4method{multisplit}{data.frame,formula}(data, grouping, ...)

\S4method{multisplit}{data.frame,character}(data, grouping, ...)
}
\arguments{
\item{data}{data frame containing several subsets}

\item{grouping}{either a model formula specifying dependent,
  independent and grouping variables in the form:<cr>
  \code{dependend ~ independend | group1 + group2 + ...}
or a character vector of column names defining the groups}

\item{formula}{model formula specifying dependent, independent and grouping
variables in the form:
\code{dependend ~ independend | group1 + group2 + ...}}
}
\value{
list containing the data subsets as its elements
}
\description{
A data frame is split into a list of data subsets defined by multiple groups.
}
\examples{


data(bactgrowth)

## simple method
spl <- multisplit(bactgrowth, c("strain", "conc", "replicate"))

## preferred method
spl <- multisplit(bactgrowth, value ~ time | strain + conc + replicate)

## show what is in one data set
spl[[1]]
summary(spl[[1]])

## use factor combination
spl[["D:0:1"]]
summary(spl[["D:0:1"]])


lapply(spl, FUN=function(x)
 plot(x$time, x$value,
      main=paste(x[1, "strain"], x[1, "conc"], x[1, "replicate"], sep=":")))


}
\seealso{
\code{\link{split}}
}

