% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\docType{methods}
\name{names,multiple_fits-method}
\alias{names,multiple_fits-method}
\alias{names-multiple_fits-method}
\alias{names<-,multiple_fits-method}
\title{Get Names of a Multiple Growth Models Object}
\usage{
\S4method{names}{multiple_fits}(x)

\S4method{names}{multiple_fits}(x) <- value
}
\arguments{
\item{x}{an object containing multiple growth models}

\item{value}{a character vector of up to the same length as x, or NULL}
}
\value{
character vector of the parameter names
}
\description{
Method to get or set names of a multiple growth models object
}
\details{
This function can be applied to the objects returned by
  \code{\link{all_growthmodels}}, \code{\link{all_splines}} and
  \code{\link{all_easylinear}}
}
\examples{

L <- all_splines(value ~ time | strain + conc + replicate,
       data = bactgrowth)

names(L)

}

