% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names.growthmodel}
\alias{names.growthmodel}
\title{Get Parameter Names of a Growth Model}
\usage{
\method{names}{growthmodel}(x)
}
\arguments{
\item{x}{a function that is parametric growth model in \pkg{growthmodels}}
}
\value{
character vector of the parameter names
}
\description{
Method to get the parameter names of a growth model
}
\details{
This function returns information about valid varameter names
  and allows automatic checking. For functions containing no \code{pnames} 
  attribute \code{names} returns \code{NULL}.
}
\examples{

names(grow_baranyi)

}

