% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\docType{methods}
\name{names,growthmodel-method}
\alias{names,growthmodel-method}
\alias{names,multiple_fits-method}
\alias{names-growthmodel}
\alias{names-multiple_fits}
\alias{names-multiple_fits<-}
\alias{names<-,multiple_fits-method}
\title{Get Names Attributes of Growth Models}
\usage{
\S4method{names}{growthmodel}(x)

\S4method{names}{multiple_fits}(x)

\S4method{names}{multiple_fits}(x) <- value
}
\arguments{
\item{x}{either a function being a parametric growth model of
package \pkg{growthmodels} or an object with multiple fits.}

\item{value}{a character vector of up to the same length as x, or NULL}
}
\value{
character vector of the parameter names
}
\description{
Methods to get the parameter names of a growth model or to get or set
  identifiers of \code{\link{multiple_fits}} objects.
}
\details{
The \code{\link{growthmodel}} meththod returns information about
  valid varameter names and allows automatic checking.
  For functions containing no \code{pnames} attribute \code{names} returns
  \code{NULL}.\cr
  The \code{\link{multiple_fits}} method can be applied to the objects
  returned by \code{\link{all_growthmodels}}, \code{\link{all_splines}} or
  \code{\link{all_easylinear}} respectively. This can be useful for selecting
  subsets, e.g. for plotting.
}
\examples{

## growthmodel-method
names(grow_baranyi)

## multiple_fits-method
L <- all_splines(value ~ time | strain + conc + replicate,
       data = bactgrowth)

names(L)

## plot only the 'R' strain
par(mfrow=c(4, 6))
plot(all.fits[grep("R:", names(all.fits))])

}
\seealso{
\code{\link{multiple_fits}}, \code{\link{all_growthmodels}},
  \code{\link{all_splines}}, \code{\link{all_easylinear}}
}

