% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalised-proportion.R
\name{normalise_proportion}
\alias{normalise_proportion}
\title{Calculate a normalised risk ration from proportions}
\usage{
normalise_proportion(
  modelled = i_proportion_model,
  base = i_baseline_proportion_data,
  ...
)
}
\arguments{
\item{modelled}{Model output from processing the \code{raw} dataframe with something like
\code{proportion_locfit_model}

A dataframe containing the following columns: 
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
}

No mandatory groupings.

No default value.
}

\item{base}{The baseline data must be grouped in the same way as \code{modelled}.

A dataframe containing the following columns: 
\itemize{
\item baseline_proportion (proportion) - Size of population
}

No mandatory groupings.

No default value.
}

\item{...}{not used}
}
\value{
a dataframe with incidence rates per unit capita.
A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
\item risk_ratio.0.025 (positive_double) - lower confidence limit of the excess risk ratio for a population group
\item risk_ratio.0.5 (positive_double) - median estimate of the excess risk ratio for a population group
\item risk_ratio.0.975 (positive_double) - upper confidence limit of the excess risk ratio for a population group
\item baseline_proportion (proportion) - The population baseline risk from which the excess risk ratio is based
}

No mandatory groupings.

No default value.
}
\description{
This assumes case distribution proportions are stratified by a population grouping, e.g.
geography or age, and we have estimates of the size of that population during
that time period. Normalising by population proportion allows us to compare groups.
}
\examples{



tmp = growthrates::england_covid \%>\%
  growthrates::proportion_locfit_model(window=21) \%>\%
  growthrates::normalise_proportion(growthrates::england_demographics) \%>\%
  dplyr::glimpse()

plot_growth_phase(tmp)

}
\concept{models}
