% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth-phase.R
\name{plot_growth_phase}
\alias{plot_growth_phase}
\title{Plot an incidence or proportion vs. growth phase diagram}
\usage{
plot_growth_phase(
  modelled = i_timestamped,
  timepoints = NULL,
  duration = max(dplyr::count(modelled)$n),
  interval = 7,
  mapping = if (interfacer::is_col_present(modelled, class)) ggplot2::aes(colour = class)
    else ggplot2::aes(),
  cis = TRUE,
  ...
)
}
\arguments{
\item{modelled}{Either:

A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

No mandatory groupings.

No default value.

OR:

A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
\item relative.growth.fit (double) - an estimate of the relative growth rate
\item relative.growth.se.fit (double) - the standard error the relative growth rate
\item relative.growth.0.025 (double) - lower confidence limit of the relative growth rate
\item relative.growth.0.5 (double) - median estimate of the relative growth rate
\item relative.growth.0.975 (double) - upper confidence limit of the relative growth rate
}

No mandatory groupings.

No default value.}

\item{timepoints}{timepoints (as \code{Date} or \code{time_period} vector) of dates to
plot phase diagrams. If multiple this will result in a sequence of plots
as facets. If \code{NULL} (the default) it will be the last time point in the series}

\item{duration}{the length of the growth rate phase trail}

\item{interval}{the length of time between markers on the phase plot}

\item{mapping}{a \code{ggplot2::aes()} mapping}

\item{cis}{should the phases be marked with confidence intervals?}

\item{...}{
  Arguments passed on to \code{\link[=geom_events]{geom_events}}
  \describe{
    \item{\code{events}}{Significant events or time spans

A dataframe containing the following columns: 
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

No mandatory groupings.

A default value is defined.
}
  }}
}
\value{
a ggplot timeseries
}
\description{
Plot an incidence or proportion vs. growth phase diagram
}
\examples{
# example code
tmp = growthrates::england_covid \%>\%
  time_aggregate(count=sum(count))

tmp_pop = growthrates::england_demographics \%>\%
  dplyr::ungroup() \%>\%
  dplyr::summarise(population = sum(population))



# If the incidence is normalised by population
tmp2 = tmp \%>\%
  poisson_locfit_model() \%>\%
  normalise_incidence(tmp_pop)

timepoints = as.Date(c("Lockdown 1" = "2020-03-30", "Lockdown 2" = "2020-12-31"))

plot_growth_phase(tmp2, timepoints, duration=108)
}
\concept{vis}
