% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-incidence-ts.R
\name{plot_incidence}
\alias{plot_incidence}
\title{Plot an incidence timeseries}
\usage{
plot_incidence(
  modelled = i_incidence_model,
  raw = i_incidence_data,
  ...,
  mapping = if (interfacer::is_col_present(modelled, class)) ggplot2::aes(colour = class)
    else ggplot2::aes(),
  events = i_events
)
}
\arguments{
\item{modelled}{An optional estimate of the incidence time series. If \code{modelled} is
missing then it is estimated from \code{raw} using a \code{poisson_locfit_model}.
In this case parameters \code{window} and \code{deg} may be supplied to control the
fit.

A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
}

No mandatory groupings.

No default value.

\code{modelled} can also be the output from \code{normalise_incidence} in which case
the plot uses the per capita rates calculated by that function}

\item{raw}{A dataframe containing the following columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified timeframe
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
}

No mandatory groupings.

No default value.}

\item{...}{
  Arguments passed on to \code{\link[=geom_events]{geom_events}}, \code{\link[=poisson_locfit_model]{poisson_locfit_model}}
  \describe{
    \item{\code{window}}{a number of data points defining the bandwidth of the estimate,
smaller values result in less smoothing, large value in more. The default
value of 14 is calibrated for data provided on a daily frequency, with
weekly data a lower value may be preferred.}
    \item{\code{deg}}{polynomial degree (min 1) - higher degree results in less
smoothing, lower values result in more smoothing. A degree of 1 is fitting
a linear model piece wise.}
    \item{\code{frequency}}{the density of the output estimates as a time period such as
\verb{7 days} or \verb{2 weeks}.}
  }}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence timeseries in the plot}

\item{events}{A dataframe containing the following columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

No mandatory groupings.

A default value is defined.}
}
\value{
a ggplot object
}
\description{
Plot an incidence timeseries
}
\examples{
# example code
tmp = growthrates::england_covid \%>\%
  time_aggregate(count=sum(count))

tmp_pop = growthrates::england_demographics \%>\%
  dplyr::ungroup() \%>\%
  dplyr::summarise(population = sum(population))

# If the incidence is normalised by population
tmp2 = tmp \%>\%
  poisson_locfit_model() \%>\%
  normalise_incidence(tmp_pop)

plot_incidence(tmp2,tmp \%>\% dplyr::cross_join(tmp_pop),colour="blue",size=0.25)
}
