% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-multinomial.R
\name{plot_multinomial}
\alias{plot_multinomial}
\title{Plot a mulitnomial proportions mode}
\usage{
plot_multinomial(
  modelled = i_multinomial_proportion_model,
  ...,
  mapping = ggplot2::aes(fill = class),
  events = i_events,
  normalise = FALSE
)
}
\arguments{
\item{modelled}{A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item class (factor) - A factor specifying the type of observation. This will be things like
}

variant, or serotype, for a multinomial model. Any missing data points are ignored.
\itemize{
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
}

Must be grouped by: class (exactly).

No default value.}

\item{...}{
  Arguments passed on to \code{\link[=geom_events]{geom_events}}
  \describe{
    \item{\code{}}{}
  }}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence timeseries in the plot}

\item{events}{A dataframe containing the following columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

No mandatory groupings.

A default value is defined.}

\item{normalise}{make sure the probabilities add up to one - this can be a
bad idea if you know you may have missing values.}
}
\value{
a ggplot
}
\description{
Plot a mulitnomial proportions mode
}
\examples{
tmp = growthrates::england_covid \%>\%
  growthrates::proportion_locfit_model(window=21) \%>\%
  dplyr::glimpse()

plot_multinomial(tmp, normalise=TRUE)+
  ggplot2::scale_fill_viridis_d()

}
