% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-proportion-ts.R
\name{plot_proportion}
\alias{plot_proportion}
\title{Plot a proportions timeseries}
\usage{
plot_proportion(
  modelled = i_proportion_model,
  raw = i_proportion_data,
  ...,
  mapping = if (interfacer::is_col_present(modelled, class)) ggplot2::aes(colour = class)
    else ggplot2::aes(),
  events = i_events
)
}
\arguments{
\item{modelled}{A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
}

No mandatory groupings.

No default value.}

\item{raw}{A dataframe containing the following columns:
\itemize{
\item denom (positive_integer) - Total test counts associated with the specified timeframe
\item count (positive_integer) - Positive case counts associated with the specified timeframe
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
}

No mandatory groupings.

No default value.}

\item{...}{
  Arguments passed on to \code{\link[=geom_events]{geom_events}}, \code{\link[=proportion_locfit_model]{proportion_locfit_model}}
  \describe{
    \item{\code{window}}{a number of data points defining the bandwidth of the estimate,
smaller values result in less smoothing, large value in more. The default
value of 14 is calibrated for data provided on a daily frequency, with
weekly data a lower value may be preferred.}
    \item{\code{deg}}{polynomial degree (min 1) - higher degree results in less
smoothing, lower values result in more smoothing. A degree of 1 is fitting
a linear model piece wise.}
    \item{\code{frequency}}{the density of the output estimates as a time period such as
\verb{7 days} or \verb{2 weeks}.}
  }}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence timeseries in the plot}

\item{events}{A dataframe containing the following columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

No mandatory groupings.

A default value is defined.}
}
\value{
a ggplot object
}
\description{
Plot a proportions timeseries
}
\examples{
tmp = growthrates::england_covid \%>\%
  growthrates::proportion_locfit_model(window=21) \%>\%
  dplyr::glimpse()

plot_proportion(tmp)+
  ggplot2::scale_fill_viridis_d(aesthetics = c("fill","colour"))
}
\concept{vis}
