% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rt.R
\name{plot_rt}
\alias{plot_rt}
\title{Reproduction number timeseries diagram}
\usage{
plot_rt(
  modelled = i_reproduction_number,
  ...,
  mapping = if (interfacer::is_col_present(modelled, class)) ggplot2::aes(colour = class)
    else ggplot2::aes(),
  events = i_events
)
}
\arguments{
\item{modelled}{A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item rt.fit (double) - an estimate of the reproduction number
\item rt.se.fit (double) - the standard error of the reproduction number
\item rt.0.025 (double) - lower confidence limit of the reproduction number
\item rt.0.5 (double) - median estimate of the reproduction number
\item rt.0.975 (double) - upper confidence limit of the reproduction number
}

No mandatory groupings.

No default value.}

\item{...}{
  Arguments passed on to \code{\link[=geom_events]{geom_events}}
  \describe{
    \item{\code{}}{}
  }}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence time series in the plot}

\item{events}{A dataframe containing the following columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

No mandatory groupings.

A default value is defined.}
}
\value{
a ggplot timeseries
}
\description{
Reproduction number timeseries diagram
}
\examples{
# example code
tmp = growthrates::england_covid \%>\%
  time_aggregate(count=sum(count))
if (FALSE) {

  tmp2 = tmp \%>\%
    poisson_locfit_model() \%>\%
    rt_from_growth_rate()

  # comparing RT from growth rates with England consensus Rt:
  plot_rt(tmp2,colour="blue")+
    geom_errorbar(data=england_consensus_rt, mapping=aes(x=date-21,ymin=low,ymax=high),colour="red")

}
}
