% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-estimators.R
\name{poisson_glm_model}
\alias{poisson_glm_model}
\title{Poisson time-series model.}
\usage{
poisson_glm_model(d, ..., window = 14, frequency = "1 day", predict = TRUE)
}
\arguments{
\item{d}{input dataframe with: \code{time} (as a \code{time_period}), \code{count} columns.
This must be ungrouped (but can be combined with a \code{group_modify})}

\item{...}{not used and present to allow proportion model to be used ina  group_modify}

\item{window}{a number of data points between knots, smaller values result in
less smoothing, large value in more.}

\item{frequency}{the density of the output estimates.}
}
\value{
a new dataframe with time (as a time period)
}
\description{
Poisson time-series model.
}
\examples{
data4 \%>\%
  mutate(time = cut_date(date, unit = "1 week", anchor="start", output="time_period") \%>\%
  group_by(time, serotype) \%>\%
  summarise(count = n(), .groups="drop_last") \%>\%
  group_by(serotype) \%>\%
  group_modify(poisson_model, df=4) \%>\%
  glimpse()
}
