% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-estimators.R
\name{proportion_glm_model}
\alias{proportion_glm_model}
\title{Binomial time-series model.}
\usage{
proportion_glm_model(d, ..., window = 14, frequency = "1 day")
}
\arguments{
\item{d}{input dataframe with: \code{time} (as a \code{time_period}), \code{count}, and \code{denom} columns.
This must be ungrouped (but can be combined with a \code{group_modify})}

\item{...}{not used and present to allow proportion model to be used ina group_modify}

\item{window}{a number of data points between knots, smaller values result in
less smoothing, large value in more.}

\item{frequency}{the density of the output estimates.}
}
\value{
a new dataframe with \code{time} (as a time period), and \code{proportion.0.5} plus quantiles, or a model object

a new dataframe with time (as a time period)
}
\description{
Binomial time-series model.
}
