% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{restore_groups}
\alias{restore_groups}
\title{Restore original format to a timed_df
#'}
\usage{
restore_groups(timed_df)
}
\arguments{
\item{timed_df}{the data frame to restore}
}
\value{
a copy of the dataframe with groups expanded and class renamed
}
\description{
Restore original format to a timed_df
#'
}
