% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-rt-wallinga.R
\name{rt_from_growth_rate}
\alias{rt_from_growth_rate}
\title{Wallinga-Teunis reproduction number}
\usage{
rt_from_growth_rate(
  df = i_growth_rate,
  ip = i_infectivity_profile,
  bootstraps = 20
)
}
\arguments{
\item{df}{A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

No mandatory groupings.

No default value.}

\item{ip}{A dataframe containing the following columns:
\itemize{
\item boot (anything) - a bootstrap identifier
\item time (positive_double) - the end of the time period (in days)
\item probability (proportion) - the probability of infection between previous time period until \code{time}
}

Must be grouped by: boot (exactly).

A default value is defined.}

\item{bootstraps}{\itemize{
\item the number of bootstraps to take to calculate for each point.
}}
}
\value{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item rt.fit (double) - an estimate of the reproduction number
\item rt.se.fit (double) - the standard error of the reproduction number
\item rt.0.025 (double) - lower confidence limit of the reproduction number
\item rt.0.5 (double) - median estimate of the reproduction number
\item rt.0.975 (double) - upper confidence limit of the reproduction number
}

No mandatory groupings.

No default value.
}
\description{
Calculate a reproduction number estimate from growth rate using the Wallinga
2007 estimation using empirical generation time distribution. This uses
resampling to transmit uncertainty in growth rate estimates
}
\examples{
tmp = growthrates::england_covid \%>\%
  time_aggregate(count=sum(count))


if (FALSE) {
  # not run
  tmp2 = tmp \%>\%
    poisson_locfit_model() \%>\%
    rt_from_growth_rate()
}

}
