% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-rt-incidence.R
\name{rt_from_incidence}
\alias{rt_from_incidence}
\title{Reproduction number from modelled incidence}
\usage{
rt_from_incidence(df = i_incidence_model, ip = i_infectivity_profile)
}
\arguments{
\item{df}{Count data

A dataframe containing the following columns: 
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
}

No mandatory groupings.

No default value.
}

\item{ip}{Infectivity profile

A dataframe containing the following columns: 
\itemize{
\item boot (anything) - a bootstrap identifier
\item time (positive_double) - the end of the time period (in days)
\item probability (proportion) - the probability of infection between previous time period until `time`
}

Must be grouped by: boot (exactly).

A default value is defined.
}
}
\value{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item rt.fit (double) - an estimate of the reproduction number
\item rt.se.fit (double) - the standard error of the reproduction number
\item rt.0.025 (double) - lower confidence limit of the reproduction number
\item rt.0.5 (double) - median estimate of the reproduction number
\item rt.0.975 (double) - upper confidence limit of the reproduction number
}

No mandatory groupings.

No default value.
}
\description{
Calculate a reproduction number estimate from growth rate using the methods
described in the vignette "Estimating the reproduction number from modelled
incidence" and using an empirical generation time distribution.
}
\examples{
df = growthrates::england_covid \%>\%
  time_aggregate(count=sum(count)) \%>\%
    poisson_locfit_model()


if (FALSE) {
  # not run
  tmp2 = df \%>\% rt_from_incidence()
}

}
\concept{models}
