% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{scale_y_log1p}
\alias{scale_y_log1p}
\title{A log1p y scale}
\usage{
scale_y_log1p(..., n = 5, base = 10, dp = 0)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}

\item{n}{the number of major breaks}

\item{base}{the base for the logarithm}

\item{dp}{decimal points}
}
\value{
a ggplot scale
}
\description{
A log1p y scale
}
