% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{scale_y_logit}
\alias{scale_y_logit}
\title{A logit y scale}
\usage{
scale_y_logit(...)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}
}
\value{
a ggplot scale
}
\description{
A logit y scale
}
