% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{time_decompose}
\alias{time_decompose}
\title{Decompose a timeseries with counts to timed data at individual level}
\usage{
time_decompose(timeseries, ...)
}
\arguments{
\item{timeseries}{a timeseries with a count column}

\item{...}{not used for supporting group_modify}
}
\value{
a timed_df with a single entry for every value of count in timeseries
}
\description{
Decompose a timeseries with counts to timed data at individual level
}
