% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-rt-wallinga.R
\name{wallinga_lipsitch}
\alias{wallinga_lipsitch}
\title{Calculate the reproduction number from a growth rate estimate and an infectivity profile}
\usage{
wallinga_lipsitch(r, y, a = 1:length(y))
}
\arguments{
\item{r}{a growth rate (may be a vector)}

\item{y}{an empirical infectivity profile as a probability vector, starting at \code{P(0<t,a[1])}}

\item{a}{the end time of the estimate (defaults to single days).}
}
\value{
a reproduction number estimate based on \code{r}
}
\description{
Calculate the reproduction number from a growth rate estimate and an infectivity profile
}
\examples{
wallinga_lipsitch(r=seq(-0.1,0.1,length.out=9), y=dgamma(1:50, 5,2))
}
