% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_growthmodels.R
\name{all_growthmodels}
\alias{all_growthmodels}
\title{Fit Nonlinear Growth Models to Data Frame}
\usage{
all_growthmodels(FUN, p, df, criteria, time = "time", y = "value",
  lower = -Inf, upper = Inf, which = names(p), method = "Marq",
  transform = c("none", "log"), ..., ncores = detectCores(logical = FALSE))
}
\arguments{
\item{FUN}{function of growth model to be fitted}

\item{p}{named vector of start parameters and initial values of the growth model}

\item{df}{data frame of observational data}

\item{criteria}{vector of criteria defining subsets in the data frame}

\item{time}{character vectors with name independent variable}

\item{y}{character vector with name of dependent variable}

\item{lower}{lower bound of the parameter vector}

\item{upper}{upper bound of the parameter vector}

\item{which}{vector of parameter names that are to be fitted}

\item{method}{character vector specifying the optimization algorithm}

\item{transform}{fit model to non-transformed or log-transformed data}

\item{ncores}{number of CPU cores used for parallel computation. The number
of real cores is detected automatically by default,
but fort debugging purposes it could be wise to set \code{ncores = 1}.
Usage of logical cores does not speed up computation.}

\item{\dots}{additional parameters passed to the optimizer}
}
\value{
object containing the parameters of all fits
}
\description{
Determine maximum growth rates by nonlinear fits for
a series of experiments.
}
\examples{

\dontrun{
data(bactgrowth)
splitted.data <- multisplit(bactgrowth, c("strain", "conc", "replicate"))

## show which experiments are in splitted.data
names(dat)

## get table from single experiment
dat <- splitted.data[["D:0:1"]]

fit0 <- fit_spline(dat$time, dat$value)

fit1 <- all_splines(bactgrowth, criteria=c("strain", "conc", "replicate"),
  spar=0.5)

## initial parameters
p <- c(coef(fit0), K = max(dat$value))

## avoid negative parameters
lower = c(y0=0, mumax=0, K=0)

## fit all models
fit2 <- all_growthmodels(FUN=grow_logistic, p=p, df=bactgrowth,
  lower = lower,
  criteria = c("strain", "conc", "replicate"), ncores = 1)

results1 <- results(fit1)
results2 <- results(fit2)

plot(results1$mumax, results2$mumax, xlab="smooth splines", ylab="logistic")
}

}
\seealso{
Other fitting functions: \code{\link{all_easylinear}},
  \code{\link{all_splines}}, \code{\link{fit_easylinear}},
  \code{\link{fit_growthmodel}}, \code{\link{fit_spline}}
}

