% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_consensus_growth_rate}
\alias{england_consensus_growth_rate}
\title{The england_consensus_growth_rate dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item date (date) - the date column
\item low (numeric) - the low column
\item high (numeric) - the high column
}

No mandatory groupings.

No default value.

111 rows and 3 columns
}
\usage{
data(england_consensus_growth_rate)
}
\description{
The england_consensus_growth_rate dataset
}
\keyword{datasets}
