% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_consensus_rt}
\alias{england_consensus_rt}
\title{The SPI-M-O England consensus reproduction number}
\format{
A dataframe containing the following columns:
\itemize{
\item date (date) - the date of the estimate
\item low (numeric) - the lower published estimate of the reproduction number
\item high (numeric) - the higher published estimate of the reproduction number
}

No mandatory groupings.

No default value.

113 rows and 3 columns
}
\usage{
data(england_consensus_rt)
}
\description{
SPI-M-O used a range of different statistical and mechanistic models to
produce estimates of the  reproduction number of the epidemic from various data
sources.
}
\concept{datasets}
\keyword{datasets}
