% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_covid_pcr_positivity}
\alias{england_covid_pcr_positivity}
\title{The england_covid_pcr_positivity dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item date (date) - the date column
\item time (as.time_period) - the time column
\item count (numeric) - the count column
\item denom (numeric) - the denom column
}

No mandatory groupings.

No default value.

1413 rows and 4 columns
}
\usage{
data(england_covid_pcr_positivity)
}
\description{
The england_covid_pcr_positivity dataset
}
\keyword{datasets}
