% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_events}
\alias{england_events}
\title{Key dated in the COVID-19 response in England}
\format{
A dataframe containing the following columns:
\itemize{
\item label (character) - the label column
\item start (date) - the start column
\item end (date) - the end column
}

No mandatory groupings.

No default value.

13 rows and 3 columns
}
\usage{
data(england_events)
}
\description{
This includes mainly the dates of lockdowns, releases from social distancing
measures and the dates new variants were first detected.
}
\concept{datasets}
\keyword{datasets}
