% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_growthmodel-class.R
\docType{class}
\name{function_growthmodel-class}
\alias{function_growthmodel-class}
\title{Union Class of Growth Model or Function}
\description{
This class union comprises parametric model functions from class
\code{growthmodel} and ordinary functions to describe time-dependent
growth of organisms.
}
\seealso{
the constructor function \code{\link{growthmodel}} how to create
  instances of class \code{growthmodel}.
}
