% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aad_set_generics.R
\name{rsquared}
\alias{rsquared}
\alias{obs}
\alias{results}
\title{Additional Generic Functions}
\usage{
rsquared(object, ...)

obs(object, ...)

results(object, ...)
}
\arguments{
\item{object}{name of a 'growthrate' object}

\item{\dots}{other arguments passed to the methods}
}
\description{
These functions are specifically defined for package \pkg{growthrates},
all other generics are imported.
}
\keyword{internal}
