% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-common.R
\name{geom_events}
\alias{geom_events}
\title{Add time series event markers to a timeseries plot.}
\usage{
geom_events(
  events = i_events,
  event_label_size = 7,
  event_label_y = Inf,
  event_label_colour = "black",
  event_line_colour = "grey50",
  event_fill_colour = "grey50",
  ...
)
}
\arguments{
\item{events}{A dataframe containing the following columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

No mandatory groupings.

A default value is defined.}

\item{event_label_size}{how big to make the event label}

\item{event_label_y}{where to place the event label (0 or Inf usually)}

\item{event_label_colour}{the event label colour}

\item{event_line_colour}{the event line colour}

\item{event_fill_colour}{the event area fill}

\item{...}{not used}
}
\value{
a set of geoms for a timeseries.
}
\description{
The x axis must be a date.
}
