% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow_huang.R
\name{grow_huang}
\alias{grow_huang}
\title{Growth Model According to Huang}
\usage{
grow_huang(time, parms)
}
\arguments{
\item{time}{vector of time steps (independend variable)}

\item{parms}{named parameter vector of Huang's growth model with:
\itemize{
  \item \code{y0} initial value of population measure
  \item \code{mumax} maximum growth rate (1/time)
  \item \code{K} carrying capacity (max. total concentration of cells)
  \item \code{alpha} shape parameter determining the curvature
  \item \code{lambda} parameter determining the lag time

}}
}
\value{
vector of dependend variable (\code{y}) and its log-transformed
  values (\code{log_y}).
}
\description{
Huangs growth model written as analytical solution of the differential equations.
}
\details{
In contrast to the original publication, all measures of population abundance
(y, y0, K) are given as untransformed values. They are not log-transformed.
}
\examples{
time <- seq(0, 30, length=200)
y    <- grow_huang(time, c(y0=0.01, mumax=.1, K=0.1, alpha=1.5, lambda=3))[,"y"]
plot(time, y, type="l")
plot(time, y, type="l", log="y")
}
\references{
Huang, Lihan (2011) A new mechanistic growth model for simultaneous
determination of lag phase duration and exponential growth rate and a new
Belehdradek-type model for evaluating the effect of temperature on growth rate.
Food Microbiology , 2011, 28, 770 - 776

Huang, Lihan (2013) Introduction to USDA Integrated Pathogen Modeling
Program (IPMP). Residue Chemistry and Predictive Microbiology Research
Unit. USDA Agricultural Research Service.
http://www.ars.usda.gov/SP2UserFiles/Place/80720500/IPMP_tutorial%201-9-14.pdf
}
\seealso{
Other growth models: \code{\link{grow_baranyi}};
  \code{\link{grow_exponential}};
  \code{\link{grow_genlogistic}},
  \code{\link{ode_genlogistic}};
  \code{\link{grow_gompertz}}; \code{\link{grow_logistic}};
  \code{\link{grow_richards}}; \code{\link{grow_twostep}},
  \code{\link{ode_twostep}}
}

