% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow_twostep.R
\name{ode_twostep}
\alias{grow_twostep}
\alias{grow_twostep.R}
\alias{ode_twostep}
\title{Twostep Growth Model}
\usage{
ode_twostep(time, y, parms, ...)

grow_twostep.R(time, parms, ...)

grow_twostep(time, parms, ...)
}
\arguments{
\item{time}{vector of simulation time steps}

\item{y}{named vector with initial values of the system (yi, ya: concentration of inactive resp. active cells)}

\item{parms}{parameters of the growth model
\itemize{
   \item \code{kw} activation (``wakeup'') constant (1/time)
   \item \code{mu}  intrinsic growth rate (1/time)
   \item \code{K}  carrying capacity (max. total concentration of cells)
}}

\item{\dots}{placeholder for additional parameters (for user-extended versions of this function)}
}
\value{
matrix of class deSolve containing the simulation outputs

\itemize{
\item \code{time} time of the simulation
\item \code{yi} concentration of inactive cells
\item \code{ya} concentration of active cells
\item \code{y} total cell concentration
\item \code{log_y} natural log of total cell concentration
}
}
\description{
System of two differential equations describing bacterial growth as two-step process of activation (or adaptation) and growth.
}
\details{
Function \code{ode_twostep} is the system of differential equations, whereas
  \code{grow_twostep} runs a numerical simulation over time.
}
\examples{
time <- seq(0, 30, length=200)
parms <- c(kw = 0.1,	mu=0.2, K=0.1)
y0    <-  c(yi=0.01, ya=0.0)
out   <- ode(y0, time, ode_twostep, parms)

plot(out)

o <- grow_twostep(0:100, c(yi=0.01, ya=0.0, kw = 0.1,	mu=0.2, K=0.1))
plot(o)
}
\seealso{
Other growth models: \code{\link{grow_exponential}};
  \code{\link{grow_genlogistic}},
  \code{\link{grow_genlogistic.R}},
  \code{\link{ode_genlogistic}};
  \code{\link{grow_gompertz}}; \code{\link{grow_huang}};
  \code{\link{grow_logistic}}; \code{\link{grow_richards}}
}

