% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aac_classes.R
\docType{class}
\name{growthrates_fit-class}
\alias{growthrates_fit-class}
\alias{nonlinear_fit-class}
\alias{easylinear_fit-class}
\alias{smooth.spline_fit-class}
\alias{multiple_fits-class}
\alias{multiple_fits}
\alias{multiple_easylinear_fits-class}
\alias{multiple_nonlinear_fits-class}
\alias{multiple_smooth.spline_fits-class}
\title{S4 Classes of Package \pkg{growthrates}}
\description{
\code{growthrates_fit}: top-level class representing a growthrates fit with
any method.
}
\section{Slots}{

\describe{
\item{\code{FUN}}{model function used.}

\item{\code{fit}}{results of the model fit.}

\item{\code{obs}}{observation data used for model fitting.}

\item{\code{rsquared}}{coefficient of determination.}

\item{\code{par}}{parameters of the fit.}

\item{\code{ndx}}{index values of the time points used (for \code{easylinear_fit}).}

\item{\code{xy}}{x and y values at the maximum of the spline.}
}}

