% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-estimators.R
\name{multinomial_nnet_model}
\alias{multinomial_nnet_model}
\title{Multinomial time-series model.}
\usage{
multinomial_nnet_model(
  d,
  ...,
  window = 14,
  frequency = "1 day",
  predict = TRUE
)
}
\arguments{
\item{d}{input dataframe with: \code{time} (as a \code{time_period}), \code{class}, \code{count} columns.
This must be ungrouped (but can be combined with a \code{group_modify})}

\item{...}{not used and present to allow proportion model to be used in a group_modify}

\item{window}{a number of data points between knots, smaller values result in
less smoothing, large value in more.}

\item{frequency}{the density of the output estimates.}

\item{predict}{result a prediction. If false we return the model.}
}
\value{
a new dataframe with \code{time} (as a time period), \code{class}, and \code{proportion.0.5}, or a model object
}
\description{
Takes a list of times, classes and counts, e.g. a COGUK variant like data set
with time, (multinomial) class (e.g. variant) and count being the count in
that time period. Fits a quadratic B-spline on time to the proportion of the
data using \code{nnet::multinom}.
}
