% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multisplit.R
\name{multisplit}
\alias{multisplit}
\title{Split Data Frame According to Multiple Criteria}
\usage{
multisplit(x, criteria)
}
\arguments{
\item{x}{data frame containing several subsets}

\item{criteria}{vector of criteria defining subsets in the data frame}
}
\value{
list containing the data subsets as its elements
}
\description{
A data frame is split into a list of data subsets defined by multiple criteria.
}
\examples{
data(bactgrowth)
spl <- multisplit(bactgrowth, c("strain", "conc", "replicate"))

## show what is in one data set
spl[[1]]
summary(spl[[1]])

lapply(spl, FUN=function(x)
  plot(x$time, x$value,
       main=paste(x[1, "strain"], x[1, "conc"], x[1, "replicate"], sep=":")))
}
\seealso{
\code{\link{split}}
}

