% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locfit-estimators.R
\name{poisson_locfit_model}
\alias{poisson_locfit_model}
\title{Poisson time-series model.}
\usage{
poisson_locfit_model(
  d,
  ...,
  window = 14,
  deg = 2,
  frequency = "1 day",
  predict = TRUE
)
}
\arguments{
\item{d}{input dataframe with: \code{time} (as a \code{time_period}), \code{count} columns.
This must be ungrouped (but can be combined with a \code{group_modify})}

\item{...}{not used and present to allow proportion model to be used ina  group_modify}

\item{window}{a number of data points between knots, smaller values result in
less smoothing, large value in more.}

\item{deg}{polynomial degree}

\item{frequency}{the density of the output estimates.}

\item{predict}{result a prediction. If false we return the model.}
}
\value{
a new dataframe with time (as a time period)
}
\description{
Poisson time-series model.
}
\examples{
# 100 weeks from 2020-01-01

tmp = as.time_period(0:100, 7, "2020-01-01")
as.Date(tmp)

range(tmp)
min(tmp)
tmp2 = as.integer(as.Date(tmp))
# testthat::expect_true(all(na.omit(tmp2-lag(tmp2)) == 7))

tmp2 = as.time_period(0:23, 1/24, "2020-01-01")
as.POSIXct(tmp2)

# convert timeseries to new "unit"
tmp = as.time_period(0:100, 7, "2020-01-01")
tmp2 = as.time_period(tmp,1)
testthat::expect_equal(as.numeric(tmp2), 0:100*7)
}
