% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-locfit.R
\name{poisson_locfit_model}
\alias{poisson_locfit_model}
\title{Poisson time-series model.}
\usage{
poisson_locfit_model(
  d = i_incidence_input,
  ...,
  window = 14,
  deg = 1,
  frequency = "1 day",
  predict = TRUE
)
}
\arguments{
\item{d}{A dataframe containing the following columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified timeframe
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
}

Ungrouped.

No default value.}

\item{...}{not used and present to allow proportion model to be used in a
\code{group_modify}}

\item{window}{a number of data points defining the bandwidth of the estimate,
smaller values result in less smoothing, large value in more. The default
value of 14 is calibrated for data provided on a daily frequency, with
weekly data a lower value may be preferred.}

\item{deg}{polynomial degree (min 1) - higher degree results in less
smoothing, lower values result in more smoothing. A degree of 1 is fitting
a linear model piece wise.}

\item{frequency}{the density of the output estimates as a time period such as
\verb{7 days} or \verb{2 weeks}.}

\item{predict}{result is a prediction dataframe. If false we return the
\code{locfit} models (advanced).}
}
\value{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

No mandatory groupings.

No default value.
}
\description{
Takes a list of times and counts and fits a quasi-poisson model
fitted with a log link function to count data using local regression
using the package \code{locfit}.
}
\details{
This results is an incidence rate estimate plus an absolute exponential
growth rate estimate both based on the time unit of the input data (e.g. for
daily data the rate will be cases per day and the growth rate will be daily).
}
\examples{
growthrates::england_covid \%>\%
  growthrates::poisson_locfit_model(window=21) \%>\%
  dplyr::glimpse()

}
