% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite-estimators.R
\name{time_summarise}
\alias{time_summarise}
\title{Summarise data from a line list to a time-series of counts.}
\usage{
time_summarise(
  df,
  unit,
  anchor = "start",
  rectangular = FALSE,
  ...,
  .fill = list(count = 0)
)
}
\arguments{
\item{df}{a line list of data you want to summarise, optionally grouped.
If this is grouped then each group is treated independently. The remaining
columns must contain a \code{date} column and may contain a \code{class} column.}

\item{unit}{a period e.g. "1 week"}

\item{anchor}{one of a date, "start" or "end" or a weekday name e.g. "mon"
this will always be one of the start of the time periods we are cutting
into}

\item{rectangular}{should the resulting time series be the same length for
all groups. This is only the case if you can be sure that all data has been
provided from all otherwise missing data will be treated as zero counts.}

\item{...}{a spec for a dplyr::summary(...) - optional, and if not provided a
\code{count = dplyr::n()} or a \code{count = sum(count)} is performed.}

\item{.fill}{a list similar to tidyr::complete for values to fill
variables with}
}
\value{
a timeseries dataframe grouped by time period,
}
\description{
Summarises by group, class and time/date.
}
